# SoulSyncApp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: SoulSync API
  version: 1.0.0
paths:
...
  /api/billing/create-checkout-session:
    post:
      summary: Create Stripe Checkout Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Checkout URL
        '400':
          description: Bad request
        '409':
          description: Duplicate (idempotency)
  /api/billing/webhook:
    post:
      summary: Stripe webhook
      responses:
        '200':
          description: OK
        '400':
          description: Signature invalid
...